wallet works with BigInteger

+++ Src/Valkyrie/Runtime/Meta/Inventory/Currency.cs
@@ -1,7 +1,9 @@
+using System.Numerics;
+
 namespace Meta.Inventory
 {
     public class Currency : BaseInventoryItem
     {
-        public long Amount { get; set; }
+        public BigInteger Amount { get; set; }
     }
 }
